Make PsfFlux variance computation consistent with old version
Because we're not using the per-pixel variance in the least squares problem here, it's not clear how we should compute the error on the flux: prior to this commit, the new PsfFlux computed the flux err as the error we would have gotten if we *had* used the per-pixel variances in computing the flux in a least squares sense.  The original algorithm (and the new PsfFlux after this commit) computes the error by interpreting the measurement as a weighted aperture.