forcedPhotCcd: allow iteration over tracts when not explicit (DM-715)
We want to iterate over all tracts spanned by the data when no
tract is provided (the user doesn't want to have to look up
what tracts are spanned).  This patch determines the set of
tracts each visit touches and adds a data reference with those
tracts for each of the visit components.  This was originally
implemented for meas_mosaic, but moved here to be of more
general use, and applied to forcedPhotCcd.

Conflicts:
	python/lsst/pipe/tasks/forcedPhotCcd.py

This is effectively a cherry-pick of HSC-715's commit:
https://github.com/HyperSuprime-Cam/pipe_tasks/commit/d0e5a1a07a70e7e802df4d286906893bf4787ff4