Fix handling of algorithm metadata in forced measurement
Algorithm metadata (which goes in the header of the measurement catalog) is not
the same as task metadata (which is saved separately).  We also need the driver
task to have access to the algorithm metadata so it can be included in the
getSchemaCatalogs() implementation.  This also makes the forced measurement
tasks better parallel their single-frame counterparts.