Fix doxygen warnings due to bad links, move doxy tag for ProcessForcedImageTask
In most cases, fixing the doxygen warnings involved removing an explicit link or ref
tag, and relying on Doxygen to automatically link something it recognizes.  This still
isn't happening in the single-package Doxygen build, but it might work with the full-
stack Doxygen build.  Worse case scenario is the internal anchor name appears in
the generated HTML and no link is made, and that's the same as what happened with
the explicit link/ref commands.