DM-461 adds a new Exception class, FatalAlgorithmError in meas base
This is a newly created exception which is known by meas_base to be "FATAL".
Which means that if it is thrown, the measurement loop on both sources and exposures
will be terminated.  As per Jim, the NO_PSF flag has been removed from various
algorithms, and replaced with a throw of this new exception.

Some unit tests had to be altered to account for the change in behavior.

Remove check for noPsf flag from this test.