Fixups to SFM test to adapt to changes in slot definition.
The previous state of this test seemed a little incorrect - the task was initialized
twice, but only used once - probably due to a merge conflict resolution problem.
But that didn't cause a failure until the preceding commit that moves slot definition
from the run() method of the task to the ctor.