Add special handling for centroiders in SafeCentroidExtractor
Unlike other algorithms, Centroiders shouldn't consider it a failure if we can't feed
them anything better than the Peak, because that's how they'd need to be run if
they were the first one.  And we need to check that they aren't the slot centroider
before we try to make an alias back to the slot centroider.