Fix confusing comments about Psfs in SdssCentroid.
This fully removes legacy code that at one point substituted an arbitrary
double-Gaussian Psf when one was not provided.  It was later determined
that it was better to just fail, requiring the user to provide their own
initial-guess Psf in advance or use a different centroider.  However, this
code had previously just been commented out, and some confusing code
comments related to it had not been removed.