Replace getFixedMomentsFlux free func with SdssShape static method
This gets SdssShapeImpl out of the interface that GaussianFlux depends on,
bringing us one step closer to getting rid of it.

This also addresses a problem with GaussianFlux overestimating its errors;
instead of relying on the calc_fisher routine in SdssShape for that, we
now estimate it directly from propagation of errors for flux only.  This
is a difference in philosophy as well as implementation, as calc_fisher
attempted to propagate uncertainty in the moments into the uncertainty
in the flux, but from GaussianFlux's perspective the moments are held
fixed.