PixelFlags: add CLIPPED mask to those that are calculated for footprints
~~~Note~~~
Port of HSC-1166 safe coadd clipping. This commit is in a non-working not
fully ported state, as changes from HSC-1202 will supersede some of the
needed changes to get the port working. As the functionality will be
removed from the next port, it was not worth time getting a fully ported
stable code running, to simply delete it.