Remove "careful" option in SdssShape getWeights()
Prior to this change the "careful" parameter had no effect, which
appeared to be a bug.  After fixing this, the results - as tested on
the stack demo package - were clearly worse, in the sense that objects
that previously had reasonable moments had moments that were smaller
than the PSF.  So, instead, I'm just removing that option entirely,
which preserves the old behavior while removing the faulty logic.