Remove usage of SdssShapeImpl in SdssShape implementation.
I haven't cleaned this up as much as I could have (by e.g. moving
the code getAdaptiveMoments directly into
SdssShape::computeAdaptiveMoments), because I felt preserving
the connection to past versions of the code was more important.
Ultimately, we'll be replacing SdssShape with a new-from-scratch
adaptive moments algorithm, so I'm not concerned about the
long-term implications of the messy code here.