Rewrite test simulation code for more flexibility
This is a major refactoring of the test data
generation code in the tests.py module, which in
turn required major changes to most of the algorithm
tests.  The good news is that these are now
more concise, more complete, and in at least some
cases, more strict.

In the process, testSingleFrameMeasurement.py
was replaced with testNoiseReplacer.py (as noise
replacement was the only thing begin tested there
that wasn't already being tested elsewhere).  In
addition, testForced.py was replaced with new tests
for forced measurement in the test files for the
flux algorithms.