forcedPhotImage: allow schema to be specified in constructor
When these Tasks are run in their own stand-alone scripts, the schema
is loaded through the butler.  However, when they are run as part of a
common operation, the Tasks are instantiated before the schemas are
written out, and so they need to be given the schemas at construction
time.  This change makes both possible: the schema is used if provided,
otherwise loaded from the butler.

Note: this is essentially a cherry-pick from HSC-1063 except that
we still refer to Coadd_src_schema instead of Coadd_ref_schema.
This will likely change once we have the multiband processing
working on the LSST stack.