Move BaseMeasurementTask into its own module
Move BasePlugin, BasePluginConfig, BaseMeasurementTask and
BaseMeasurementConfig to a new module baseMeasurement.py
in order to allow BaseMeasurementTask access to ApplyApCorrTask.
Without this refactoring it would be very difficult to avoid a
circular import. This change is arguably an improvement because it
moves higher-level code out of base.py and moves BaseMeasurementTask
into a more obvious module (one that analogous to sfm.py and
forcedMeasurement.py)