Filter reference sources in forced photometry by deblend family
The ReplaceWithNoise procedure in measurement requires all child sources
to have their parents present, so we can't do any kind of filtering that
could cut a parent but not one or more of its children.

This is essentially a cherry-pick of the changes to reference.py
from HSC-1218's commit (i.e. the changes to forcedPhotImage.py were
not made as they are not appropriate):
https://github.com/HyperSuprime-Cam/pipe_tasks/commit/c442bc91cb59bf1b7143c423270afebe3c90684f