Raise error if reference list contains parent-less children
Check that the reference catalog does not contain any children for which
any member of their parent chain is not within the list.  This can occur at
boundaries when the parent is outside and one of the children is within.
Currently, the parent chain is always only one deep, but just in case, this
code checks for any case where when the parent chain to a child's topmost
parent is broken and raises an exception if it occurs.