ApplyApCorrTask.run: fix test for models in apCorrMap
Fixed the test that determined if the required models for flux
and fluxSigma were present in the aperture correction map.
The test incorrectly assumed that ApCorrMap.get raised an
exception if a field was missing; in fact it returns None.
Tested using a slightly hacked version of
meas_extensions_photometryKron tests/Kron.py
since we don't yet have a unit test for appling ap. corr.