Simplify PerTractCcdDataIdContainer inheritance
Modify PerTractCcdDataIdContainerto inherit from lsst.pipe.base.DataIdContainer
instead of CoaddDataIdContainer.  Inheriting from the latter was confusing as
it pertains to Coadds and this method pertains to Ccds.  Additionally, the only
functionality it was gaining was the setting and memoising of the skymap via
the getSkymap() function.  The memoizing was in fact also being done in the
function itself, so simply set the skymap explicitly there (thus omiting the
need to call getSkymap() and its inheritance from from CoaddDataIdContainer).