Move PerTractCcdDataIdContainer to forcedPhotCcd.py
Currently, forcedPhotCcd.py is the only place where the
PerTractCcdDataIdContainer is being used, so it seems appropriate
just to include it there (as opposed to having in its own file).
The only other place currently planned to make use of it is in
meas_mosaic, which is not yet implemented.  If a more suitable
location for this container becomes apparent (e.g. in the process
of getting meas_mosaic running with the stack), it can be moved
accordingly then.