Drop base_Variance centroid-outside-Footprint test.
The centroid falling outside the Footprint may be indicative of a deeper
problem with the centroider or the Footprint itself, but the appropriate place
to catch that error is here.

The base_Variance plugin now checks the centroid flag, and assumes its safe to
proceed if that's not set. Otherwise, we throw a MeasurementError which
(silently) sets a flag.