SingleFrameVariancePlugin: measure variance only within aperture
SingleFrameVariancePlugin.measure was taking the median of the
entire image, rather than within an aperture around the source of
interest. A Footprint was constructed with the aperture, but was
unused. Fixed by using a HeavyFootprint to get the pixels within
the aperture.

Fixing the aperture revealed a problem in the test: all the pixels
in the aperture were NAN. Found that this was because the source
was too narrow (has width=2 in the original version in HSC, but was
somehow changed to 1 here); with a wider source, the aperture is
larger and includes pixels that are not NAN. Added unmasked NAN
pixels to test that the median is not being taken from the entire
image (fails previously, passes now).