Port blendedness metrics from HSC.
This was moved out of meas_deblender and into a measurement algorithm.  However, this
plugin is different in that it needs access to the deblended pixels and the full
image without noise replacement.  Because of this, it does not implement the measure
method as other algorithms.  It currently has the plugin call the C++ functions directly.
This should be updated once the plugin framework has a more generic interface.