Ensure positive-definite shape determinant
The GaussianFlux plugin throws runtime errors when the SdssShape
plugin shape has non positive-definite determinant. SafeShapeExtractor
is updated so that rather than ensuring that the determinant is not
less than zero, SafeShapeExtarctor now checks that Ixx*Iyy >
(1 + epsilon)*Ixy*Ixy. The same check is made in SdssShapeResult.
For now, we set epsilon = 1.0e-6 but this a magic number. DM-5801 is
tasked with examining the this magic number value of epsilon.