Add afterburner plugins and a task to run them
Afterburner plugins are designed to run after measurement plugins,
and will only operate on data within a catalog. The plugins can either
take a single record at a time or the whole catalog by specifying the
PlugType property to be either 'single' or 'multi'. An AfterburnerTask
is also introduced to run the plugins. The list of plugins to run
must be specified in the AfterburnerConfig which is passed to the
AfterburnerTask init method. The AfterburnerTask is entered through
the run method which expects the catalog to be operated on to be passed
as an argument.