Change how modules are looked up in wrapped Control classes.
The pex_config C++ control object wrapper mechanism needs to set the
__module__ attribute of the ConfigClasses it creates, and it provides
two ways of doing this: by name and by location in the call stack.  It
turns out the latter does not work when a nested control object is
declared in the same module as its owner, because it necessitates an
absolute import of the module before while it is being imported.  The
stack-based method doesn't suffer from this problem, so we switch to
it as the default.