Redirect numpy RuntimeWarning to debug log.
SingleFrameVariancePlugin attempts to compute the median of the
variance array for pixels in the heavy footprint of a source. Some
pixels are ignored because mask flags are set indicating that the pixel
is not to be trusted. Sometimes, this results in all pixedls getting
rejected at which point numpy raises a RuntimeWarning stating that it
is trying to compute the median of an empty array. We redirect each
instance of this warning to a debug log where it can be disiplayed if
desirted but does not clutter the output otherwise.