Handle empty variance array gracefully.
SingleFrameVariancePlugin attempts to compute the median of the
variance array for pixels in the heavy footprint of a source. Some
pixels are ignored because mask flags are set indicating that the pixel
is not to be trusted. Sometimes, this results in all pixels getting
rejected at which point numpy raises a RuntimeWarning stating that it
is trying to compute the median of an empty array. We check for the
empty array and raise a MeasurementError if all the pixels are flagged.
The testVariance unittest has been modified to test the new behavior.