CentroidChecker forces consistency with footprints.
The CentroidChecker should be used by a Centroid Plugin.
It fixes centroider errors whose result is outside the
footprint.  It must be constructed during __init__,
as it adds a flag_resetToPeak to the subschema of the
centroid plugin.

The centroidChecker(measRecord) should be called after
a plugin stores it measured centroid in its measRecord.
If the centroid found lies outside the footprint, the
result will be "reset" to the footprint peak.

If maxDistFromPeak > 0 in the config of the plugin,
a check and reset is also done for centroids which fall
further than specified from the footprint peak.

When a reset if done, flag and flag_resetToPeak are set.

Unit test for CentroidChecker.py including toy plugin
to simulate errors. Also check that the plugin works
with the base Centroid algorithms.

Naive, Gaussian and SdssCentroid changed to use checker.