Remove unnecessary code to set FAILURE flag to False
I think this is leftover from an older style of setting
flags, which allowed the algorithm to set flags to True,
then throw out from anywhere.

Such failures will now be caught by the framework.
Plus, this strategy would circumvent any other code
(such as the Centroid Checker itself) which might set
a failure flag but allow the plugin to continue on.