Bug fix: ProcessImageForced -> ForcedPhotImage
Renamed multiple references to ProcessImageForcedTask/Config
to ForcedPhotImageTask/Config. The new name is more consistent
with the classes that inherit from these, and there were multiple
instances in the existing code base that assumed the new name
(mostly in comments, but also in code).

Note that these classes are only used in meas_base, because
the task is an abstract base class, so this change will not
affect any other module.