provide option to measure on undeblended image
In some cases, the deblender doesn't do a good job, or we want to
compare what's measured by our code with a different code that
doesn't do deblending; so we measure on the undeblended image.

Refactored the guts of the measurement (mainly the exception handling)
in the BaseMeasurementTask to make that easy. SingleFrameMeasurementTask
will now measure on the undeblended image using a new list of plugins
in SingleFrameMeasurementConfig.undeblended; the results will appear
in the catalog with columns named after the measurement algorithms but
with the SingleFrameMeasurementConfig.undeblendedPrefix prepended.