DM-9249 Change FlagHandler to remove the enum of flag types
The enum which formerly numbered the available failure flags is replaced with
a set of static FlagDefinitions: (e.g., static FlagDefinition const FAILURE;
for the general flag). The position of the flag in the FlagHandler list is
now the number value of the FlagDefinition (e.g., FAILURE.number).

A new class FlagDefinitionList is introduced to specify the
FlagDefinitions used to initialize the FlagHandler.
Add flags to the FlagDefinitionList using the methods
addFailureFlag() and add(name, doc). When done, use the FlagHandler
initializer addFields(schema, name, flagDefinitionList).

An optional exclusion list exclDefs is provided if some members of
the algorithms FlagDefinitionList are not desired.
The _flagHandler keeps the same numbering of flags whether or not
a flag is excluded -- the FlagDefinition.number is fixed.

It is now optional whether a General Failure flag is defined for the
flagHandler.

Remove the FlagHandler::FAILURE which assumes FAILURE is position 0.