Change FlagHandler to remove enum DM-9249
enum is replaced with direct access to FlagDefinitions via Flags struct

For MeasurementErrors and Result::flags setting, the client uses the
std::string identifier Algorithm::Flags::FlagDefinition::name

If the index of the flag is required, use Algorithm.getFlagNumber(flagName)
with the string identifier flagName = FlagDefinition.name as the argument.

Change getFlagDefinitions to a std::vector<FlagDefinition> const, which
means that a std::vector::const_iterator is used for initializing flagHandler.