Revised this change following discussion with Jim
Remove the use of iterators to initialize the FlagHandler and only permit a FlagDefinitionList to be used

Use a vector internally for FlagDefinitionList. This requires that
we not use of FlagDefinitions const & methods (always return copies).

Return the static flag declarations to the header files, in
the form of FlagDefinition const FLAGNAME; instead of an enum.

Remove the FlagHandler::FAILURE which was used to always identify
the failure flag as number 0. Static method on FlagDefinitionList
now replaces it.

Add the ability to skip some FlagDefinitions in the FlagHandler init.
They will still be in the FlagHandler _vector, but with invalid keys.