Modification to remove unnecessary information from FlagHandler
The static FlagDefinitionList now contain the information about
the available FlagDefinitions for an algorithm, and the doc and number.
The role of the FlagHandler is now to connect the FlagDefinitions
in the FlagDefinitionList to the database key. These two could probably
be merged into one object, but this change makes the decision not to
do this, but to make the function of each more clear.