DM-430: Modify measurement framework to log plugins error in individual logs
Logs named for the taskName + pluginName (e.g., measurement.base_PsfFlux)
Name is set by the task, but the name is pushed down to the plugin.
if the plugin class has "hasLogName = True" and an optional "logName"
parameter in its initialization.  If both are so, plugin.getLogName() wil
return the name of the log.

if lsst.log.Log logs are identified by name, not by the logger itself, so the
logName is enough.

C++ routines can use the BaseAlgorithm function getLogName to get the right log.

Log Measurement Errors, which were not previously logged.

Checkin seg fault problem with pybind methods on BaseAlgorithm
Add py:multiple_inheritance to SimpleAlgorithm wrapper.