Remove flagDecorator.py (DM-9866)
Documentation of changes to flagHandler in mainpage.doc were held out of
DM-9249, pending resolution of RFC-302, and are submitted here.

Changes to the model code in testPythonPlugin.py and to python plugins
which used the flagDecorator were made with DM-9249.

meas_extensions_ngmix was missed in these changes, and will be
repaired on this ticket.