Increase ease of measurement reconstruction
Many of the tools nessisary for reconstructing the state of an
exposure at the time of a measurement, and possibly re-running
the measurement already exist, but are not easy or convienent.
This commit aims to increase the ease of accomplishing these
tasks by introducing convienence functionality, and adding a
document which contains runable examples.