Add required methods to dummy class
The measurement plugins registered by test_ApCorrNameSet can affect
the tests in test_PluginLogs.py because they were not set up as
real plugins. To allow test_PluginLogs to run within the same
process as test_ApCorrNameSet, we now add the fail() and
getExecutionOrder() methods to allow them to adhere to the
plugin contract.