wrappers: add factories to produce SFM/forced plugins from generic
Some measurement plugins are really generic in the sense that it
doesn't matter whether they are run as a single-frame or forced
plugin, but the plumbing is slightly different (most notably, the
signature of the 'measure' method). To work around this, added a
GenericPlugin class and methods that will use that to produce the
appropriate SingleFramePlugin and ForcedPlugin classes so common
code can be used for both.