TestDataset: make deterministic by default
TestDataset contains two methods (makePerturbedWcs and realize) that
take an optional seed for the random number generator. In both cases,
the seed defaults to None, which means the RNG is seeded in some
non-deterministic manner. However, this is intended for tests, which we
require are deterministic. The current API therefore makes it too easy
(simply neglecting to specify the optional randomSeed keyword argument)
to do something bad. Instead, default to a preset (and therefore
deterministic) value for the seed. If the user wants things to change
between invocations, he can specify different seeds.