Replace Calib->PhotoCalib
PhotoCalib always just "does the math", whether it results in NaN or not.
For negative instFlux, this can result in NaN magnitude but finite magnitude
error, which can be surprising, but is the correct numerical result. This
was special cased in Calib (return NaN,Nan for negative flux), so a test here
had to be updated.