Update source ID generation to avoid the wrong magic number.
This should have been included on DM-17689, but the code here was on a
branch that was in progress at the same time and got merged later.

Some (but not all) of this fixup commit then got orphaned on a
DM-29729 branch, because the main purpose of that ticket was something
that got marked invalid.