Add parameter to skip check of plugin-slot validity
Many tests were written assuming that you could create a Measurement
task that did only some plugins, while getting values from slots that
had values in them (instead of running the associated plugins). The
config validation assumes that every slot will have an associated plugin
to run to fill it in. This is a stopgap until we come up with a better
way to check that the necessary slots are defined for all the plugins
(possibly in afw.table).