Move centroid finitness check and rework exception
Catching NaN centroids early means that the offimage flag is not set
for them, but it should be.
Reformat exception message to be more useful (although the measurement
task squashes all these to debug level logs, so noone ever sees them
anyway).
Add tests of non-finite centroids and sources right on the image boundary.