Use 'pvi', not 'initial_pvi', in forcedPhotCcd.
We also turn off using the finalVisitSummary input, which is important
for the PhotoCalib, since both pvi and initial_pvi have their pixel
values calibrated to nJy and their PhotoCalib reset to the identity.
This means forcedPhotCcd's instFlux columns will actually have nJy
units, but this is fine since their "DN" is already an arbitrary unit
and the LocalPhotoCalib column will reflect this by being full of ones.