Skip measuring sources with certain pixel flags by specifying plugin order.
This approach is inspired by ip_diffim's dipoleFitTask.
We have to run the centroid plugin first, and then we run PixelFlags next
so we can drop sources from the measCat before other plugins are run.
By default, only sources with the nodataCenter flag are dropped.
In practice, most NO_DATA regions are from template holes in diffims,
so the result will be an identical length mergedForcedSource table but
with NaNs for the fluxes of the obviously garbage diaSources.